Flask_websocket_chat/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.23.1
blinker==1.7.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
Flask==3.0.2
Flask-Login==0.6.3
flask-redis==0.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
redis==5.0.3
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.28
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
Werkzeug==3.0.1
wsproto==1.2.0
zope.interface==6.2
